fix<BE>: MemberJoinRequest의 Favorite List 자료형 Integer -> Long 변경

See merge request s07-webmobile1-sub2/S07P12B204!109

+++ backend/src/main/java/com/ssafy/cheertogether/member/dto/MemberJoinRequest.java
@@ -12,8 +12,8 @@ public class MemberJoinRequest {
 	private String profileImage;
 	private String role;
 	private String myInfo;
-	private List<Integer> favoriteLeagueList;
-	private List<Integer> favoriteTeamList;
+	private List<Long> favoriteLeagueList;
+	private List<Long> favoriteTeamList;
 
 	public void saveProfileImgLink(String link) {
 		profileImage = link;
